requirements.txt
pandas
awscli
apache-airflow
apache-airflow-providers-amazon